requirements.txt
sentence_transformers == 3.3.0
numpy == 2.1.3
openpyxl == 3.1.5
faiss-cpu  == 1.9.0  # CPU平台默认
networkx  == 3.4.2
PyMuPDF
gradio
